global: move validation functions to its file
